Move views from right to left before closing the dired pane
In a 2 column setup and with views in the second column, switching to a
single layout reverses the tab order of what we had in the right column,
changes the focused/top view of that group, and also looses the MRU
order (`ctrl+tab`).

Mitigate this behavior by moving and focusing before changing the
layout.  We still loose the MRU info.