Unconditionally remove `DiredMoveOpenOrNewFileToRightGroup`
The code intends to move files from a dired pane to the other pane.
That doesn't work as there is no correct detection of the dired pane
and the other pane.  E.g. a dired pane should probably *only* has
exactly one dired view (likely to mimic the sidebar), but that's not
tested here, and the comment also indicates that it should still be
possible to have multiple views in such a dired panel.

Just remove that as it moves files around if you have a dired browser
open in group 0 which is obviously a bug.

When tracing the code origin, it seems this is added to move files to
the correct position when opening them while being in the dired browser.
But that has to be done just there, imperatively in the commands.