Tune paste behavior
dired used to paste into directories under the cursor.

This is inconvenient as a user might copy some files to the buffer,
`<backspace>` to go up, and `ctrl+v` immediately. This did not work.
Instead the user had to navigate away from the directory section at the
top, down to the files section.  (Or alternatively to the "..".)

Change that and treat collapsed directories just like files.  Keep the
old behavior for expanded files.